rename files according to class name
